Move DataTab components into a /dataTab/ folder
