+++ src/components/dataTab/DataTabClassButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { UIMode } from "../App";
+import { UIMode } from "../../App";
 import { makeStyles } from "@material-ui/core";
-import FormButton from "./FormButton";
+import FormButton from "../FormButton";
 
 const useStyles = makeStyles({
   classSelectorButtonArea: {
