+++ src/components/dataTab/DataTabOperatorButton.tsx
@@ -1,10 +1,10 @@
 ﻿import React from "react";
 import slugify from "slugify";
-import { Operator } from "../App";
+import { Operator } from "../../App";
 import { makeStyles } from "@material-ui/core";
-import { useRarityStyles } from "./StyleRarityUnderline";
+import { useRarityStyles } from "../StyleRarityUnderline";
 import clsx from "clsx";
-import FormButton from "./FormButton";
+import FormButton from "../FormButton";
 
 const useStyles = makeStyles({
   iconStack: {
